Added changes to EplNmtk, Edrv82573, EdrvCyclic, EplDllk, ...
- EplApiGeneric: Fixed compiler warning
- demo_mn_82573_kernel: Updated project
- EplNmtk: Added error handling for EplDllkProcess
- Edrv82573: Increased number of tx buffers
- EdrvCyclic: Added new error code for case when no cycle length is set
- EplDllk: Added handling of WaitSoCPReq delay

+++ EplStack/EplApiGeneric.c
@@ -267,7 +267,7 @@ tEplDllkInitParam   DllkInitParam;
     // reset instance structure
     EPL_MEMSET(&EplApiInstance_g, 0, sizeof (EplApiInstance_g));
 
-    EPL_MEMCPY(&EplApiInstance_g.m_InitParam, pInitParam_p, min(sizeof (tEplApiInitParam), pInitParam_p->m_uiSizeOfStruct));
+    EPL_MEMCPY(&EplApiInstance_g.m_InitParam, pInitParam_p, min(sizeof (tEplApiInitParam), (size_t) pInitParam_p->m_uiSizeOfStruct));
 
     // check event callback function pointer
     if (EplApiInstance_g.m_InitParam.m_pfnCbEvent == NULL)
